PIM-9715: refactoring bis
