+++ tests/back/Pim/Enrichment/Integration/EventSubscriber/UpdateFamilyAttributeAsLabelIntegration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace AkeneoTest\Pim\Structure\Integration\Family;
+namespace AkeneoTest\Pim\Enrichment\Integration\EventSubscriber;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
