+++ tests/back/Pim/Enrichment/Specification/Bundle/EventSubscriber/Family/UpdateFamilyAttributeAsLabelSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Structure\Bundle\EventSubscriber;
+namespace Specification\Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Family;
 
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\Repository\AttributeRepositoryInterface;
